Moving go2lua internal package to regbot
