+++ cmd/regbot/internal/go2lua/export.go
