+++ cmd/regbot/internal/go2lua/go2lua_test.go
