+++ cmd/regbot/internal/go2lua/import.go
